Use special characters for the character column
I personally do not like overloading '.' for both '.' and non-printable
characters.  Also, I don't like that NUL is usually displayed as '.' as
well.

So let me take the opportunity of Unicode terminals and use '·' for
non-printable characters, an empty space ' ' for NUL, and '␣' for actual
(0x20) spaces.  Oh, and '¶' for \n, because this is my program.