Make command line editing not impossible
Supporting backspace is easy, so we can do that (and we really should).
Supporting the rest is harder, so let us put it off (but at least do
something useful, that is, replace the \e by ^[ so the user can see the
damage that has been done and easily revert it, instead of making the
terminal do magic).