Fix escape sequence overflow
When doing a lot of scrolls, it was possible to exhaust the
escape_sequence buffer.  Then, some sequence was cut in half and treated
as proper normal input.

Fix this mainly by aborting escape sequence processing if a second \e
occurs, but also by stuffing the tail after some escape sequence back
into the input, and by increasing the escape_sequence buffer size
(because I guess 32 may not be enough for all weird mouse things, I
 don't know).