Update buffer when aborting replacement
when a replacement is aborted mid-way (that is, the user replaced the
first nibble but not the second), the file is not modified.  We should
update the buffer (that is, reload its contents from the file) to
reflect that, or the user will think the first nibble has indeed been
modified.